Implement basic functions + Created Drawio Class Diagram
